Hotgym: separate HelloSPTPTest into cpp, hpp, and test
as gtest cannot live with a main()
