+++ src/examples/Hotgym.cpp
@@ -1,0 +1,8 @@
+#include "examples/HelloSPTP.cpp"
+
+//this runs as executable
+int main() {
+  using namespace examples;
+  run();
+  return 0;
+}
